Rename MutantExecutionResultFactory to TestFrameworkMutantExecutionResultFactory and add interface
Extracted from https://github.com/infection/infection/pull/2098 to simplify a big PR

Why?

Because for PHPUnit process and for Static Analysis process (PHPStan in particular) we can not use the same ExecutionResult factory, because

- processes exit with different status codes
- PHPStan uses `stderr` for a lot of "internal" things, we can't rely on it
- because we don't need to parse `stdout` for "passed tests" in case of PHPStan
- and so on

So, each MutantProcess inside MutantProcessContainer (see https://github.com/infection/infection/pull/2096) will have different execution result parsers.
