+++ devTools/phpstan-baseline.neon
@@ -37,6 +37,12 @@ parameters:
 			path: ../src/Logger/Html/StrykerHtmlReportBuilder.php
 
 		-
+			message: '#^Unused Infection\\Mutant\\MutantExecutionResultFactory\:\:createFromProcess$#'
+			identifier: shipmonk.deadMethod
+			count: 1
+			path: ../src/Mutant/MutantExecutionResultFactory.php
+
+		-
 			message: '#^Parameter \#1 \$array of function array_intersect expects an array of values castable to string, list\<PhpParser\\Node\\Expr\|string\> given\.$#'
 			identifier: argument.type
 			count: 1
