+++ src/Mutant/MutantExecutionResultFactory.php
@@ -35,91 +35,12 @@ declare(strict_types=1);
 
 namespace Infection\Mutant;
 
-use Infection\AbstractTestFramework\SyntaxErrorAware;
-use Infection\AbstractTestFramework\TestFrameworkAdapter;
 use Infection\Process\MutantProcess;
-use function sprintf;
-use Symfony\Component\Process\Process;
-use Webmozart\Assert\Assert;
 
 /**
  * @internal
- * @final
  */
-class MutantExecutionResultFactory
+interface MutantExecutionResultFactory
 {
-    private const PROCESS_MIN_ERROR_CODE = 100;
-
-    public function __construct(
-        private readonly TestFrameworkAdapter $testFrameworkAdapter,
-    ) {
-    }
-
-    public function createFromProcess(MutantProcess $mutantProcess): MutantExecutionResult
-    {
-        $process = $mutantProcess->getProcess();
-        $mutant = $mutantProcess->getMutant();
-        $mutation = $mutant->getMutation();
-
-        return new MutantExecutionResult(
-            $process->getCommandLine(),
-            $this->retrieveProcessOutput($process),
-            $this->retrieveDetectionStatus($mutantProcess),
-            $mutant->getDiff(),
-            $mutation->getHash(),
-            $mutation->getMutatorClass(),
-            $mutation->getMutatorName(),
-            $mutation->getOriginalFilePath(),
-            $mutation->getOriginalStartingLine(),
-            $mutation->getOriginalEndingLine(),
-            $mutation->getOriginalStartFilePosition(),
-            $mutation->getOriginalEndFilePosition(),
-            $mutant->getPrettyPrintedOriginalCode(),
-            $mutant->getMutatedCode(),
-            $mutant->getTests(),
-        );
-    }
-
-    private function retrieveProcessOutput(Process $process): string
-    {
-        Assert::true(
-            $process->isTerminated(),
-            sprintf(
-                'Cannot retrieve a non-terminated process output. Got "%s"',
-                $process->getStatus(),
-            ),
-        );
-
-        return $process->getOutput();
-    }
-
-    private function retrieveDetectionStatus(MutantProcess $mutantProcess): string
-    {
-        if (!$mutantProcess->getMutant()->isCoveredByTest()) {
-            return DetectionStatus::NOT_COVERED;
-        }
-
-        if ($mutantProcess->isTimedOut()) {
-            return DetectionStatus::TIMED_OUT;
-        }
-
-        $process = $mutantProcess->getProcess();
-
-        if ($process->getExitCode() > self::PROCESS_MIN_ERROR_CODE) {
-            // See \Symfony\Component\Process\Process::$exitCodes
-            return DetectionStatus::ERROR;
-        }
-
-        $output = $this->retrieveProcessOutput($process);
-
-        if ($process->getExitCode() === 0 && $this->testFrameworkAdapter->testsPass($output)) {
-            return DetectionStatus::ESCAPED;
-        }
-
-        if ($this->testFrameworkAdapter instanceof SyntaxErrorAware && $this->testFrameworkAdapter->isSyntaxError($output)) {
-            return DetectionStatus::SYNTAX_ERROR;
-        }
-
-        return DetectionStatus::KILLED;
-    }
+    public function createFromProcess(MutantProcess $mutantProcess): MutantExecutionResult;
 }
